requirements/base.txt
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
# Django
# ------------------------------------------------------------------------------
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
# from pillar project
# ------------------------------------------------------------------------------------
aenum==3.1.0
boto3==1.18.6
celery==5.1.2
certifi==2021.5.30
cryptography==3.4.7
Django==3.2.5
djangorestframework==3.12.4
django-environ==0.4.5
django-ckeditor==6.1.0
django-clear-cache==0.3
django-cors-headers==3.7.0
django-health-check==3.16.4
django-imagekit==4.0.2
django-mptt==0.12.0
django-push-notifications==2.0.0
django-redis==5.0.0
django-rest-framework==0.1.0 # don't remove
django-rest-swagger==2.2.0 # don't remove
django-rq==2.4.1
django-silk==4.1.0
django-storages==1.11.1
fitparse==1.2.0
kombu==5.1.0
openpyxl==3.0.7  # For reading xlsx file
Pillow==8.3.1 # needed, but don't know why
psycopg2-binary==2.9.1
pycparser==2.20
python-json-logger==2.0.1
python3-openid==3.2.0
pytz==2021.1
readwise-django-rq-scheduler==1.2.1
requests==2.26.0
requests-oauthlib==1.3.0
sentry-sdk==1.3.0
six==1.16.0
urllib3==1.26.6
stravalib==0.10.2
newrelic==6.4.4.161
drf-yasg==1.20.0
gevent==21.1.2 # for gunicorn threading
XlsxWriter==1.4.4  # For writing in a xlsx file
smart_open[s3]==5.1.0  # For opening xlsx file
uWSGI==2.0.19.1
django-log-request-id==2.0.0
